Avoid the old service injection API
Using Ember.inject.service is what's suggested, so I think it makes more sense to just stick with that.  Also, for some reason by injecting the ApplicationController into the tour service in the initializer (so we can reference the currentRoute property), the ApplicationController wasn't able to use `this.tour` to reference the tour, despite the `Ember.inject`-based API working fine.